Test: 구독하기api 테스트코드작성

+++ src/channel/channel.service.ts
@@ -52,13 +52,10 @@ export class ChannelService {
   }
 
   async createSubscribe(userIdx: number, channelIdx: number): Promise<void> {
-    console.log('실행');
     const subscribeState = await this.getSubscrbeState(userIdx, channelIdx);
-    console.log(subscribeState);
     if (subscribeState) {
       throw new ConflictException('already in subscribe');
     }
-    console.log('실행1');
 
     await this.prisma.subscribe.create({
       data: { subscriber: userIdx, provider: channelIdx },
@@ -69,6 +66,11 @@ export class ChannelService {
     userIdx: number,
     channelIdx: number,
   ): Promise<void> {
+    const subscribeState = await this.getSubscrbeState(userIdx, channelIdx);
+    if (!subscribeState) {
+      throw new ConflictException('Already not subscribe');
+    }
+
     await this.prisma.subscribe.deleteMany({
       where: { subscriber: userIdx, provider: channelIdx },
     });
